percpu refactor
