+++ .vscode/settings.json
@@ -18,7 +18,7 @@
         // "anemone-kernel/crates/buddy-system/fuzz/Cargo.toml",
     ],
     "rust-analyzer.cargo.target": "riscv64gc-unknown-none-elf",
-    //"rust-analyzer.cargo.target": "loongarch64-unknown-none",
+    // "rust-analyzer.cargo.target": "loongarch64-unknown-none",
     "rust-analyzer.cargo.allTargets": false,
     "files.associations": {
         "conf/.defconfig": "toml",
