+++ anemone-kernel/src/driver/intc/sifive_plic.rs
@@ -206,7 +206,7 @@ impl CoreIrqChip for SiFivePlic {
 
                 let regs = plic.regs();
 
-                for i in 0..CpuArch::ncpus() {
+                for i in 0..ncpus() {
                     let ctx = SiFivePlic::s_ctx_for(i);
 
                     // keep all sources masked by default, individual lines are unmasked by
@@ -230,9 +230,7 @@ impl CoreIrqChip for SiFivePlic {
     }
 
     fn claim(&self) -> Option<HwIrq> {
-        let claimed =
-            self.regs()
-                .claim(SiFivePlic::s_ctx_for(CpuArch::cur_cpu_id().get())) as usize;
+        let claimed = self.regs().claim(SiFivePlic::s_ctx_for(cur_cpu_id().get())) as usize;
         if claimed == 0 {
             return None;
         }
@@ -264,6 +262,6 @@ impl SiFivePlic {
     }
 
     fn current_s_context(&self) -> usize {
-        Self::s_ctx_for(CpuArch::cur_cpu_id().get())
+        Self::s_ctx_for(cur_cpu_id().get())
     }
 }
