+++ anemone-kernel/src/main.rs
@@ -32,6 +32,7 @@ pub mod fs;
 pub mod initcall;
 pub mod mm;
 pub mod panic;
+pub mod percpu;
 pub mod power;
 pub mod sched;
 pub mod sync;
@@ -51,6 +52,7 @@ use crate::{
         probe_virtual_devices,
     },
     mm::layout::KernelLayoutTrait,
+    percpu::percpu_login,
     prelude::*,
     sync::{counter::CpuSync, mono::MonoOnce},
     task::{
@@ -176,7 +178,7 @@ unsafe extern "C" fn bsp_kinit(bsp_id: usize, fdt_va: VirtAddr) {
 
     #[cfg(feature = "kunit")]
     {
-        crate::debug::kunit::kunit_runner();
+        // crate::debug::kunit::kunit_runner();
         unsafe {
             KUNIT_SYNC_COUNTER.sync_with_counter();
         }
